fix: free unused buffers (home) in strings.c
the function didn't free the buffer
this frees the buffer on error by passing the array to bmalloc()

+++ src/sync/main.c
@@ -57,7 +57,7 @@ int main(int argc, char *argv[])
 
 	char *home = NULL;	/* will be allocated soon (maybe) */
 	void *buf_arr[3] = { NULL, NULL, NULL };
-	uint8_t buf_i = 0;
+	uint16_t buf_i = 0;
 
 	if (use_home_dir_src)
 	{
